config: Extract `BlockConfig` struct
This moves the `blocked_traffic`, `blocked_ips` and `blocked_routes`
settings from `Server` onto a dedicated `BlockConfig` struct in
`config::block`, along with the traffic pattern parsing helpers and their
tests. The fields are renamed to `traffic`, `ips` and `routes` since the
struct now provides the `block` namespace.
